ref(autofix): split green check suite into a package
Break green_check_suite.py into base/ready_for_review/review_request with a
shared GreenCheckSuiteContext, so each side effect owns its own marker,
lock and metrics. base only gates and dispatches.

Also stamp every skip/failure with the suite's conclusion_type by
replacing record_check_suite_skip with record_check_suite_outcome, which
records both skips and failures per stage.
